Update writeup with the full set of bisect results
The results for all languages of performing the regression best-fits
have been updated to use non-normalised data (in practice the weightings
have the same effect, but the fitting works more effectively).

The data for applying the weighted distance metric bisect algorithm has
also been added in, in the form of means and standard deviations for how
quickly the regression commit was discovered using the algorithm.

Eight new graphs for this last set of results covering C and JavaScript
projects have also been added.